refacto study flow components
