--- src/components/study/perimeter/StudyFlow.module.css
@@ -1,16 +1,0 @@
-.flowButtons {
-  gap: 1rem;
-}
-
-.flowUploadButton {
-  display: none;
-}
-
-.flowImage {
-  max-width: 100%;
-  max-height: 30rem;
-}
-
-.error {
-  color: red;
-}
