+++ src/css/variables.css
@@ -17,4 +17,6 @@
   --primary-80: #272768;
 
   --header-height: 4rem;
+
+  --error-00: #ff0000;
 }
