+++ src/i18n/fr.json
@@ -250,17 +250,7 @@
       }
     },
     "perimeter": {
-      "addFlow": "Ajouter un flux",
       "dates": "Du {startDate} au {endDate}",
-      "documentNotFound": "Document non trouvé",
-      "flows": "Cartographie des fluxs",
-      "flowSelector": "Flux à afficher",
-      "invalidFileType": "Type de fichier non supporté",
-      "newFlow": "Ajouter un flux",
-      "noFileSelected": "Aucun fichier selectionné",
-      "noFlows": "Aucun flux à afficher",
-      "notAuthorized": "Action non authorisée",
-      "removeFlow": "Supprimer le flux",
       "title": "Périmètre de l'étude : {name}"
     },
     "results": {
@@ -296,6 +286,20 @@
       "uncertainty": "Incertitude",
       "validation": "Validation de la source d'émission",
       "yes": "Oui"
+    },
+
+    "flow": {
+      "addFlow": "Ajouter un flux",
+      "documentNotFound": "Document non trouvé",
+      "flows": "Cartographie des fluxs",
+      "flowSelector": "Flux à afficher",
+      "invalidFileType": "Type de fichier non supporté",
+      "newFlow": "Ajouter un flux",
+      "noFileSelected": "Aucun fichier selectionné",
+      "noFlows": "Aucun flux à afficher",
+      "notAuthorized": "Action non authorisée",
+      "removeFlow": "Supprimer le flux",
+      "Not authorized": "Vous n'avez pas les droits pour effectuer cette action"
     }
   },
   "emissionFactors": {
