+++ src/services/serverFunctions/file.ts
@@ -1,13 +1,12 @@
 'use server'
 
 import { Document } from '@prisma/client'
-import { getFileUrlFromBucket } from '../file'
-import { NOT_AUTHORIZED } from '../permissions/check'
 import { canAccessFlowFromStudy } from '../permissions/study'
+import { getFileUrlFromBucket } from '../serverFunctions/scaleway'
 
 export const getDocumentUrl = async (document: Document, studyId: string) => {
   if (!(await canAccessFlowFromStudy(document.id, studyId))) {
-    return NOT_AUTHORIZED
+    return ''
   }
   return getFileUrlFromBucket(document.bucketKey)
 }
